fix: Remove styling class.
