+++ src/PixelDust.Game/GUI/Elements/PGUIElement.cs
@@ -1,8 +1,14 @@
-﻿using Microsoft.Xna.Framework;
+﻿using System.Collections.Generic;
+
+using Microsoft.Xna.Framework;
 
 using PixelDust.Game.Objects;
 
-using System.Collections.Generic;
+using PixelDust.Game.Enums.General;
+using PixelDust.Game.Enums.GUI;
+using PixelDust.Game.Mathematics;
+using PixelDust.Game.Constants;
+
 
 namespace PixelDust.Game.GUI.Elements
 {
@@ -11,49 +17,95 @@ namespace PixelDust.Game.GUI.Elements
         // Settings
         public string Id { get; set; }
 
-        // System
-        protected PGUILayout GUILayout { get; private set; }
-
         // Readonly
-        public PGUIElementStyle Style => this.style;
         public Vector2 Position => this.position;
 
         // Parental
         public PGUIElement Parent => this.parent;
         public PGUIElement[] Children => [.. this.children];
         public bool HasChildren => this.children.Count > 0;
-
-        private readonly PGUIElementStyle style;
+        public PPositioningType PositioningType => this.positioningType;
+        public PCardinalDirection PositionAnchor => this.positionAnchor;
+        public Size2 Size => this.size;
+        public Vector2 Margin => this.margin;
 
         private readonly List<PGUIElement> children = [];
         private readonly Dictionary<string, object> data = [];
 
-        private PGUIElement parent;
+        private PPositioningType positioningType;
+        private PCardinalDirection positionAnchor;
+        private Size2 size;
+        private Vector2 margin;
         private Vector2 position;
+
+        private PGUIElement parent;
 
         public PGUIElement()
         {
-            this.style = new(this);
             this.Id = string.Empty;
         }
 
         public PGUIElement(string id)
         {
-            this.style = new(this);
             this.Id = id;
         }
 
-        internal void SetGUILayout(PGUILayout layout)
+        #region Settings
+        public Vector2 GetPosition()
         {
-            this.GUILayout = layout;
+            Size2 screenSize = new(PScreenConstants.DEFAULT_SCREEN_WIDTH, PScreenConstants.DEFAULT_SCREEN_HEIGHT);
+
+            Size2 targetSize = screenSize;
+            Vector2 targetPosition = Vector2.Zero;
+
+            if (this.positioningType == PPositioningType.Relative && this.parent != null)
+            {
+                targetSize = this.parent.size;
+                targetPosition = this.parent.position;
+            }
+
+            return this.PositionAnchor switch
+            {
+                PCardinalDirection.Center => targetPosition + this.Margin + (new Vector2(targetSize.Width, targetSize.Height) / 2),
+                PCardinalDirection.North => targetPosition + this.Margin + new Vector2(targetSize.Width / 2, 0),
+                PCardinalDirection.Northeast => targetPosition + this.Margin + new Vector2(targetSize.Width, 0),
+                PCardinalDirection.East => targetPosition + this.Margin + new Vector2(targetSize.Width, targetSize.Height / 2),
+                PCardinalDirection.Southeast => targetPosition + this.Margin + new Vector2(targetSize.Width, targetSize.Height),
+                PCardinalDirection.South => targetPosition + this.Margin + new Vector2(targetSize.Width / 2, targetSize.Height),
+                PCardinalDirection.Southwest => targetPosition + this.Margin + new Vector2(0, targetSize.Height),
+                PCardinalDirection.West => targetPosition + this.Margin + new Vector2(0, targetSize.Height / 2),
+                PCardinalDirection.Northwest => targetPosition + this.Margin,
+                _ => targetPosition,
+            };
         }
+
+        public void SetPositioningType(PPositioningType type)
+        {
+            this.positioningType = type;
+        }
+
+        public void SetPositionAnchor(PCardinalDirection cardinalDirection)
+        {
+            this.positionAnchor = cardinalDirection;
+        }
+
+        public void SetSize(Size2 size)
+        {
+            this.size = size;
+        }
+
+        public void SetMargin(Vector2 margin)
+        {
+            this.margin = margin;
+        }
+        #endregion
 
         #region Parental
         public void AppendChild(PGUIElement element)
         {
             element.parent?.RemoveChild(element);
             element.parent = this;
-            element.position = element.style.GetPosition();
+            element.position = element.GetPosition();
 
             this.children.Add(element);
         }
@@ -68,7 +120,6 @@ namespace PixelDust.Game.GUI.Elements
 
         public void RemoveChild(PGUIElement element)
         {
-            element.parent = this.GUILayout.RootElement;
             _ = this.children.Remove(element);
         }
         #endregion
